javascript carousel
i began adding the js to the document to implement a photo carousel. i
started writing the code in the head, and found that it wasn’t
inserting the item from the array into the div i created. I moved the
code from the head to the body and it worked!.